requirements.txt
cachetools==4.1.1
certifi==2020.12.5
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
croniter==0.3.36
cryptography==3.0
dask==2.30.0
distributed==2.30.1
docker==4.4.0
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-core==1.4.4
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
HeapDict==1.0.1
idna==2.10
importlib-resources==3.3.0
marshmallow==3.9.1
marshmallow-oneofschema==2.1.0
msgpack==1.0.0
mypy-extensions==0.4.3
natsort==7.1.0
numpy==1.19.4
pandas==1.1.5
pendulum==2.1.2
prefect==0.13.18
protobuf==3.14.0
psutil==5.7.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==19.1.0
python-box==5.2.0
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.4
pytzdata==2020.1
PyYAML==5.3.1
requests==2.25.0
rsa==4.6
six==1.15.0
sortedcontainers==2.3.0
tabulate==0.8.7
tblib==1.7.0
text-unidecode==1.3
toml==0.10.2
toolz==0.11.1
tornado==6.1
urllib3==1.26.2
websocket-client==0.57.0
zict==2.0.0
zipp==3.4.0
